casting

+++ components/data_reader/cams/base_cam.cpp
@@ -25,7 +25,7 @@ void data_reader::BaseCam::setCamIntrinsics(const int width, const int height, c
   _cy = static_cast<double>(_frameSize.height) / 2.0;
   // Set field of view
   _horizontalFov = horizontalFov;
-  _verticalFov = (_frameSize.height *_horizontalFov) / _frameSize.width;
+  _verticalFov = (static_cast<double>(_frameSize.height) *_horizontalFov) / static_cast<double>(_frameSize.width);
   // Set focal length
   _fx = (static_cast<double>(width) * 0.5) / tan(_horizontalFov * 0.5);
   _fx = _fy; // assuming concentric lens
